remove usage of New dashboard guardian
